+++ EliteSpeedrunTool/dataobserver/datafetcher/MissionHashFetcher.h
@@ -2,10 +2,10 @@
 
 #include "BaseDataFetcher.h"
 
-class MissionHashFetcher : public BaseDataFetcher<unsigned long long> {
+class MissionHashFetcher : public BaseDataFetcher<unsigned int> {
     Q_OBJECT
 public:
     explicit MissionHashFetcher(QObject* parent = nullptr);
 
-    unsigned long long fetchData() override;
+    unsigned int fetchData() override;
 };
